Settings: Change Subscription "Order Created" label to "Order Pending payment"
WooCommerce Orders don't have a "Created" status.  When an Order is created, it is assigned the "Pending payment" status.

This changes the integration's settings to match the same labels as used in WooCommerce Orders, and adds a more verbose description to the settings field to reflect https://woocommerce.com/document/managing-orders/

The underlying status value ("pending") is accurate, and does not need to be changed.